13 add speech to text solution analysis

+++ shopping_list.py
@@ -10,10 +10,13 @@ def change_item_bought_status(item_id, bought):
 
 with database.ShoppingListDatabase() as db:
     items = db.get_items()
+    is_first_unbought_found = False
     if items:
         sorted_items = sorted(items, key=lambda x: x[4])
         divider_drawn = False
         for item in sorted_items:
+            if not is_first_unbought_found and item[4]:
+                is_first_unbought_found = True
             item_id, item_name, quantity, unit, bought = item
             
             if bought and not divider_drawn:
@@ -33,11 +36,12 @@ with database.ShoppingListDatabase() as db:
 
 st.container(height=20, border=False)
 
-if st.button("Usuń kupione produkty", width="stretch"):
-    with database.ShoppingListDatabase() as db:
-        items = db.get_items()
-        for item in items:
-            item_id, item_name, quantity, unit, bought = item
-            if bought:
-                db.delete_item(item_id)
-    st.rerun()
+if is_first_unbought_found:
+    if st.button("Usuń kupione produkty", width="stretch"):
+        with database.ShoppingListDatabase() as db:
+            items = db.get_items()
+            for item in items:
+                item_id, item_name, quantity, unit, bought = item
+                if bought:
+                    db.delete_item(item_id)
+        st.rerun()
